[issue-3] extract parsing of config file to "fractals::parser"
